+++ src/redox/mod.rs
@@ -1,3 +1,6 @@
+use std::fs::File;
+use std::io::Read;
+
 use {Type, Info, Version};
 
 pub fn current_platform() -> Info {
@@ -5,4 +8,14 @@ pub fn current_platform() -> Info {
         os_type: Type::Redox,
         version: Version::unknown(),
     }
+}
+
+fn get_version() -> Option<String> {
+    let mut file = match File::open("sys:uname") {
+        Some(file) => file,
+        _ => return None,
+    };
+
+    let mut version = String::new();
+    file.read_to_string(&mut contents).ok().map(|_| version)
 }